Test: API 이동 테스트에 반영
Ref: #72
