+++ src/test/java/com/studypals/domain/groupManage/restDocsTest/GroupControllerRestDocsTest.java
@@ -10,8 +10,6 @@ import static org.springframework.restdocs.http.HttpDocumentation.httpResponse;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
-import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -26,7 +24,6 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.studypals.domain.groupManage.api.GroupController;
 import com.studypals.domain.groupManage.dto.*;
-import com.studypals.domain.groupManage.entity.GroupRole;
 import com.studypals.domain.groupManage.service.GroupService;
 import com.studypals.global.responses.CommonResponse;
 import com.studypals.global.responses.Response;
@@ -104,76 +101,5 @@ public class GroupControllerRestDocsTest extends RestDocsSupport {
                                         .description("그룹 가입 시 승인 필요 여부 / Default FALSE")
                                         .attributes(constraints("not null"))),
                         responseHeaders(headerWithName("Location").description("추가된 그룹 id"))));
-    }
-
-    @Test
-    @WithMockUser
-    void generateEntryCode_success() throws Exception {
-        // given
-        Long groupId = 1L;
-        GroupEntryCodeRes entryCodeRes = new GroupEntryCodeRes(groupId, "A1B2C3");
-        Response<GroupEntryCodeRes> expected = CommonResponse.success(ResponseCode.GROUP_ENTRY_CODE, entryCodeRes);
-
-        given(groupService.generateEntryCode(any(), any())).willReturn(entryCodeRes);
-
-        // when
-        ResultActions result = mockMvc.perform(post("/groups/" + groupId + "/entry-code"));
-
-        // then
-        result.andExpect(status().isCreated())
-                .andExpect(header().string("Location", "/groups/1/entry-code/" + entryCodeRes.code()))
-                .andExpect(hasKey(expected))
-                .andDo(restDocs.document(
-                        httpRequest(),
-                        httpResponse(),
-                        responseFields(
-                                fieldWithPath("data.groupId").description("그룹 ID"),
-                                fieldWithPath("data.code").description("그룹 초대 코드 | 6자리의 대문자 알파벳, 숫자 조합"),
-                                fieldWithPath("code").description("응답 코드"),
-                                fieldWithPath("status").description("응답 상태"),
-                                fieldWithPath("message").description("응답 메시지")),
-                        responseHeaders(headerWithName("Location").description("생성된 그룹 초대 코드"))));
-    }
-
-    @Test
-    void getGroupSummary_success() throws Exception {
-        // given
-        String entryCode = "A1B2C3";
-        List<GroupSummaryRes.GroupMemberProfileImageDto> profiles = List.of(
-                new GroupSummaryRes.GroupMemberProfileImageDto("imageUrl url", GroupRole.LEADER),
-                new GroupSummaryRes.GroupMemberProfileImageDto("imageUrl url", GroupRole.MEMBER));
-        GroupSummaryRes groupSummaryRes = GroupSummaryRes.builder()
-                .id(1L)
-                .name("group name")
-                .tag("tag")
-                .isOpen(true)
-                .memberCount(2)
-                .profiles(profiles)
-                .build();
-        Response<GroupSummaryRes> expected = CommonResponse.success(ResponseCode.GROUP_SUMMARY, groupSummaryRes);
-
-        given(groupService.getGroupSummary(entryCode)).willReturn(groupSummaryRes);
-
-        // when
-        ResultActions result = mockMvc.perform(get("/groups/summary").param("entryCode", entryCode));
-
-        // then
-        result.andExpect(status().isOk())
-                .andExpect(hasKey(expected))
-                .andDo(restDocs.document(
-                        httpRequest(),
-                        httpResponse(),
-                        queryParameters(parameterWithName("entryCode").description("그룹 초대 코드")),
-                        responseFields(
-                                fieldWithPath("data.id").description("그룹 ID"),
-                                fieldWithPath("data.name").description("그룹명"),
-                                fieldWithPath("data.tag").description("그룹 태그"),
-                                fieldWithPath("data.isOpen").description("그룹 공개 여부"),
-                                fieldWithPath("data.memberCount").description("그룹 전체 멤버 수"),
-                                fieldWithPath("data.profiles[].imageUrl").description("그룹 멤버 프로필 이미지"),
-                                fieldWithPath("data.profiles[].role").description("그룹 멤버 권한 | LEADER, MEMBER"),
-                                fieldWithPath("code").description("응답 코드"),
-                                fieldWithPath("status").description("응답 상태"),
-                                fieldWithPath("message").description("응답 메시지"))));
     }
 }
