+++ src/test/java/com/studypals/domain/groupManage/service/GroupEntryServiceTest.java
@@ -2,8 +2,12 @@ package com.studypals.domain.groupManage.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willThrow;
+
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,14 +15,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import com.studypals.domain.groupManage.dto.GroupEntryCodeRes;
 import com.studypals.domain.groupManage.dto.GroupEntryReq;
+import com.studypals.domain.groupManage.dto.GroupMemberProfileDto;
+import com.studypals.domain.groupManage.dto.GroupSummaryRes;
 import com.studypals.domain.groupManage.entity.Group;
 import com.studypals.domain.groupManage.entity.GroupEntryRequest;
 import com.studypals.domain.groupManage.entity.GroupMember;
-import com.studypals.domain.groupManage.worker.GroupEntryCodeManager;
-import com.studypals.domain.groupManage.worker.GroupEntryRequestWorker;
-import com.studypals.domain.groupManage.worker.GroupMemberWorker;
-import com.studypals.domain.groupManage.worker.GroupReader;
+import com.studypals.domain.groupManage.entity.GroupRole;
+import com.studypals.domain.groupManage.worker.*;
 import com.studypals.domain.memberManage.entity.Member;
 import com.studypals.domain.memberManage.worker.MemberReader;
 import com.studypals.global.exceptions.errorCode.GroupErrorCode;
@@ -43,10 +48,104 @@ public class GroupEntryServiceTest {
     private GroupEntryRequestWorker entryRequestWorker;
 
     @Mock
+    private GroupMemberReader groupMemberReader;
+
+    @Mock
+    private GroupAuthorityValidator authorityValidator;
+
+    @Mock
     private Member mockMember;
 
     @InjectMocks
     private GroupEntryServiceImpl groupEntryService;
+
+    @Test
+    void generateEntryCode_success() {
+        // given
+        Long userId = 1L;
+        Long groupId = 1L;
+        Group group = Group.builder().id(1L).build();
+        String entryCode = "A1B2C3";
+        GroupEntryCodeRes expected = new GroupEntryCodeRes(groupId, entryCode);
+
+        given(groupReader.getById(groupId)).willReturn(group);
+        given(entryCodeManager.generate(groupId)).willReturn(entryCode);
+
+        // when
+        GroupEntryCodeRes actual = groupEntryService.generateEntryCode(userId, groupId);
+
+        // then
+        assertThat(actual).isEqualTo(expected);
+    }
+
+    @Test
+    void generateEntryCode_fail_invalidAuthority() {
+        // given
+        Long userId = 1L;
+        Long groupId = 1L;
+        GroupErrorCode errorCode = GroupErrorCode.GROUP_FORBIDDEN;
+
+        willThrow(new GroupException(errorCode)).given(authorityValidator).validate(userId, groupId);
+
+        // when & then
+        assertThatThrownBy(() -> groupEntryService.generateEntryCode(userId, groupId))
+                .isInstanceOf(GroupException.class)
+                .extracting("errorCode")
+                .isEqualTo(errorCode);
+    }
+
+    @Test
+    void generateEntryCode_fail_groupNotFound() {
+        // given
+        Long userId = 1L;
+        Long groupId = 1L;
+        GroupErrorCode errorCode = GroupErrorCode.GROUP_NOT_FOUND;
+
+        given(groupReader.getById(groupId)).willThrow(new GroupException(errorCode));
+
+        // when & then
+        assertThatThrownBy(() -> groupEntryService.generateEntryCode(userId, groupId))
+                .isInstanceOf(GroupException.class)
+                .extracting("errorCode")
+                .isEqualTo(errorCode);
+    }
+
+    @Test
+    void getGroupSummary_success() {
+        // given
+        String entryCode = "entry code";
+        Group group = Group.builder()
+                .id(1L)
+                .name("group")
+                .tag("tag")
+                .isOpen(true)
+                .totalMember(5)
+                .build();
+        List<GroupMemberProfileDto> profiles = List.of(
+                new GroupMemberProfileDto(1L, "name", "imageUrl url", GroupRole.LEADER),
+                new GroupMemberProfileDto(2L, "name2", "imageUrl url", GroupRole.MEMBER));
+        GroupSummaryRes expected = GroupSummaryRes.builder()
+                .id(group.getId())
+                .name(group.getName())
+                .tag(group.getTag())
+                .isOpen(group.isOpen())
+                .memberCount(group.getTotalMember())
+                .profiles(profiles.stream()
+                        .map(it -> new GroupSummaryRes.GroupMemberProfileImageDto(it.imageUrl(), it.role()))
+                        .toList())
+                .build();
+
+        given(entryCodeManager.getGroupId(entryCode)).willReturn(group.getId());
+        given(groupReader.getById(group.getId())).willReturn(group);
+        given(groupMemberReader.getTopNMemberProfiles(eq(group.getId()), anyInt()))
+                .willReturn(profiles);
+
+        // when
+        GroupSummaryRes actual = groupEntryService.getGroupSummary(entryCode);
+
+        // then
+        assertThat(actual).isEqualTo(expected);
+    }
 
     @Test
     void joinGroup_success() {
