+++ src/test/java/com/studypals/domain/groupManage/service/GroupServiceTest.java
@@ -2,10 +2,7 @@ package com.studypals.domain.groupManage.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.willThrow;
 
 import java.util.List;
 
@@ -18,7 +15,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.studypals.domain.groupManage.dto.*;
 import com.studypals.domain.groupManage.dto.mappers.GroupMapper;
 import com.studypals.domain.groupManage.entity.Group;
-import com.studypals.domain.groupManage.entity.GroupRole;
 import com.studypals.domain.groupManage.entity.GroupTag;
 import com.studypals.domain.groupManage.worker.*;
 import com.studypals.domain.memberManage.entity.Member;
@@ -49,15 +45,6 @@ public class GroupServiceTest {
 
     @Mock
     private GroupMemberWorker groupMemberWorker;
-
-    @Mock
-    private GroupMemberReader groupMemberReader;
-
-    @Mock
-    private GroupAuthorityValidator authorityValidator;
-
-    @Mock
-    private GroupEntryCodeManager entryCodeManager;
 
     @Mock
     private GroupMapper groupMapper;
@@ -137,93 +124,5 @@ public class GroupServiceTest {
                 .isInstanceOf(GroupException.class)
                 .extracting("errorCode")
                 .isEqualTo(errorCode);
-    }
-
-    @Test
-    void generateEntryCode_success() {
-        // given
-        Long userId = 1L;
-        Long groupId = 1L;
-        String entryCode = "A1B2C3";
-        GroupEntryCodeRes expected = new GroupEntryCodeRes(groupId, entryCode);
-
-        given(mockGroup.getId()).willReturn(groupId);
-        given(groupReader.getById(groupId)).willReturn(mockGroup);
-        given(entryCodeManager.generate(groupId)).willReturn(entryCode);
-
-        // when
-        GroupEntryCodeRes actual = groupService.generateEntryCode(userId, groupId);
-
-        // then
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @Test
-    void generateEntryCode_fail_invalidAuthority() {
-        // given
-        Long userId = 1L;
-        Long groupId = 1L;
-        GroupErrorCode errorCode = GroupErrorCode.GROUP_FORBIDDEN;
-
-        willThrow(new GroupException(errorCode)).given(authorityValidator).validate(userId, groupId);
-
-        // when & then
-        assertThatThrownBy(() -> groupService.generateEntryCode(userId, groupId))
-                .isInstanceOf(GroupException.class)
-                .extracting("errorCode")
-                .isEqualTo(errorCode);
-    }
-
-    @Test
-    void generateEntryCode_fail_groupNotFound() {
-        // given
-        Long userId = 1L;
-        Long groupId = 1L;
-        GroupErrorCode errorCode = GroupErrorCode.GROUP_NOT_FOUND;
-
-        given(groupReader.getById(groupId)).willThrow(new GroupException(errorCode));
-
-        // when & then
-        assertThatThrownBy(() -> groupService.generateEntryCode(userId, groupId))
-                .isInstanceOf(GroupException.class)
-                .extracting("errorCode")
-                .isEqualTo(errorCode);
-    }
-
-    @Test
-    void getGroupSummary_success() {
-        // given
-        String entryCode = "entry code";
-        Group group = Group.builder()
-                .id(1L)
-                .name("group")
-                .tag("tag")
-                .isOpen(true)
-                .totalMember(5)
-                .build();
-        List<GroupMemberProfileDto> profiles = List.of(
-                new GroupMemberProfileDto(1L, "name", "imageUrl url", GroupRole.LEADER),
-                new GroupMemberProfileDto(2L, "name2", "imageUrl url", GroupRole.MEMBER));
-        GroupSummaryRes expected = GroupSummaryRes.builder()
-                .id(group.getId())
-                .name(group.getName())
-                .tag(group.getTag())
-                .isOpen(group.isOpen())
-                .memberCount(group.getTotalMember())
-                .profiles(profiles.stream()
-                        .map(it -> new GroupSummaryRes.GroupMemberProfileImageDto(it.imageUrl(), it.role()))
-                        .toList())
-                .build();
-
-        given(entryCodeManager.getGroupId(entryCode)).willReturn(group.getId());
-        given(groupReader.getById(group.getId())).willReturn(group);
-        given(groupMemberReader.getTopNMemberProfiles(eq(group.getId()), anyInt()))
-                .willReturn(profiles);
-
-        // when
-        GroupSummaryRes actual = groupService.getGroupSummary(entryCode);
-
-        // then
-        assertThat(actual).isEqualTo(expected);
     }
 }
